+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## [v0.0.2]
+
+- Adds a Projectile game object prefab
+- Adds Projectile.cs
+- Player spawns projectile
+
+
 ## [v0.0.1] 2025-06-08
 
 - adds Player game object